% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{box_plot}
\alias{box_plot}
\title{An ad hoc box plot function to visualize results obtained through \code{rep_simulations} and joined through \code{join_results}}
\usage{
box_plot(results_table, title)
}
\arguments{
\item{results_table}{a dataframe containing results}

\item{title}{a string representing title for the plot}
}
\value{
a ggplot
}
\description{
An ad hoc box plot function to visualize results obtained through \code{rep_simulations} and joined through \code{join_results}
}
