% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{join_results}
\alias{join_results}
\title{An ad hoc utility function to combine the results of univariate statistical testing between two different parameter settings to make it easy for plotting}
\usage{
join_results(results_list_1, results_list_2, table_name)
}
\arguments{
\item{results_list_1}{A list of results obtained through \code{rep_simulations} for one particular parameter setting}

\item{results_list_2}{A list of results obtained through \code{rep_simulations} for one particular parameter setting}

\item{table_name}{A string among "deseq_results" or "logistic_results"}
}
\value{
A dataframe containing results binded together
}
\description{
An ad hoc utility function to combine the results of univariate statistical testing between two different parameter settings to make it easy for plotting
}
