% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_single_analysis.R
\name{rep_simulations}
\alias{rep_simulations}
\title{Repeats the single analysis (run_single_analysis) desired number of times for quantifying uncertainty}
\usage{
rep_simulations(sim_params_list, n_times)
}
\arguments{
\item{sim_params_list}{A list object in R containing parameter specifications to run the analysis, where the legal parameters are described in the documentation}

\item{n_times}{An integer specifying the number of replications of the analysis}
}
\value{
A list containing results
}
\description{
Repeats the single analysis (run_single_analysis) desired number of times for quantifying uncertainty
}
