% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_analysis.R
\name{run_single_analysis}
\alias{run_single_analysis}
\title{Run a single iteration of analysis with the parameter specifications}
\usage{
run_single_analysis(sim_params)
}
\arguments{
\item{sim_params}{A list object in R containing parameter specifications to run the analysis, where the legal parameters are described in the documentation}
}
\value{
A list containing results
}
\description{
Run a single iteration of analysis with the parameter specifications
}
