% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{write_list_to_disk}
\alias{write_list_to_disk}
\title{An ad hoc utility function to write the results of analyses to disk. Writes nested lists using folder structures.}
\usage{
write_list_to_disk(mylist, filepath)
}
\arguments{
\item{mylist}{a list of analyses results}

\item{filepath}{path to write the results}
}
\value{
writes r objects to disk, but does not return anything
}
\description{
An ad hoc utility function to write the results of analyses to disk. Writes nested lists using folder structures.
}
